Move tracer classes into separate files
